pyproject.toml
[tool.poetry]
name = "pyPVCY"
version = "0.1.0"
description = "A Python API client and CLI for the Privacy Dynamics API."
authors = ["Ted Conbeer <tconbeer@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "pvcy", from="src"}]

[tool.poetry.scripts]
pvcy = 'pvcy.__main__:__main__'

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
pydantic = ">=2.0"
click = "^8.1.6"
ruamel-yaml = "^0.17.32"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.269"
pytest = "^7.3.1"
black = "^23.3.0"
pre-commit = "^3.3.2"
mypy = "^1.3.0"
types-oauthlib = "^3.2.0.8"
types-requests = "^2.31.0.0"
types-pyyaml = "^6.0.12.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["A", "B", "E", "F", "I"]
target-version = "py39"

[tool.mypy]
python_version = "3.9"
files = [
    "src/pvcy/**/*.py",
    "tests/**/*.py",
]
mypy_path = "src,tests,stubs"

show_column_numbers = true

# show error messages from unrelated files
follow_imports = "normal"

# be strict
disallow_any_unimported = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
strict_optional = true

warn_return_any = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

no_implicit_reexport = true
strict_equality = true

[tool.pytest.ini_options]
log_level = "DEBUG"
